# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Leaflow Amber",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/files/download/{hash}": {
            "get": {
                "description": "根据文件 Hash 下载文件。仅支持图片下载，且图片具有有效期",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "下载图片",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FileId uint64 `uri:\"id\" binding:\"required\"`",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}